LayoutTests/http/tests/resourceLoadStatistics/leave-empty-document-referrer-alone-third-party-iframe-expected.txt

Tests that an empty document.referrer is left alone in third-party iframes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.referrer is [empty]
PASS successfullyParsed is true

TEST COMPLETE
